command: stop restarting playback when writing to playlist-pos
0085d1abd19f675936a97115a187086f7097c8c9 added a warning about writing
the same value to the playlist-pos property that in the future it would
stop restarting playback. Instead, you should use the
playlist-play-index command for that. Well go ahead and drop the old
deprecated behavior now and do what wm4 wanted this to do: just ignore
if the same value is written again.