Default to video-latency-hacks. I'm seeing some weird playlist seeking behavior around EOF that this option seems to fix, without any downside (yet). I'm not sure how it interacts with the --demuxer-cache-wait option, which I removed a while ago to just always wait (=yes).
It only seems to happen on a few files I have, and probably only
on Windows.

I guess 18f86dc2c05e1c408802fed4baa5236b41e9eacd was supposed to
fix this. It's not exactly that seeking backward is broken, just
that it looks like it's jumping back an extra keyframe or more.